Preserve dtype in flatten_data

+++ scripts/spice/sparse.py
@@ -203,7 +203,7 @@ class SPICEBatchLoader:
         batch_num_edges = self.num_edges_tr[batch_idxs]
 
         def flatten_data(batch_data, batch_num_data, max_num_data, fill_value, offsets):
-            flattened_data = jnp.full((max_num_data, *batch_data.shape[2:]), fill_value)
+            flattened_data = jnp.full((max_num_data, *batch_data.shape[2:]), fill_value, dtype=batch_data.dtype)
             data_flattened = 0
             for i, (graph_data, graph_num_data) in enumerate(zip(batch_data, batch_num_data)):
                 next_data_idx = data_flattened + graph_num_data
